# doc/swagger.yaml
openapi: 3.0.0
paths:
...
  /category:
    post:
      operationId: CategoryController_create
      summary: 'Add new category '
      description: 'Adds new category '
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryDto'
      responses:
        '201':
          description: Category is created
          content:
            application/json:
              example:
                id: f3d2f4c6-5376-48df-b4a7-9fe825559db9
                name: string
                description: string
        '400':
          description: Bad request. Body does not contain required fields
        '403':
          description: Insufficient permissions
      tags:
        - Categories
    get:
      operationId: CategoryController_findAll
      summary: Get all categories
      description: Gets all categories
      parameters:
        - name: page
          required: false
          in: query
          description: Page number (starting from 1)
          schema:
            example: 1
            type: number
        - name: limit
          required: false
          in: query
          description: Number of items per page
          schema:
            example: 10
            type: number
        - name: sortBy
          required: false
          in: query
          description: The field to sort by
          schema:
            example: name
            enum:
              - id
              - name
              - description
            type: string
        - name: order
          required: false
          in: query
          description: Sorting direction
          schema:
            example: DESC
            enum:
              - ASC
              - DESC
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              examples:
                Without pagination:
                  value:
                    - id: f3d2f4c6-5376-48df-b4a7-9fe825559db9
                      name: string
                      description: string
                  description: Without page and limit query parameters
                With pagination:
                  value:
                    total: 1
                    page: 1
                    limit: 10
                    data:
                      - id: f3d2f4c6-5376-48df-b4a7-9fe825559db9
                        name: string
                        description: string
                  description: With page and limit query parameters
        '400':
          description: Bad request. Wrong query parameters (hacker scope)
        '403':
          description: Insufficient permissions
      tags:
        - Categories
  /category/{id}:
    get:
      operationId: CategoryController_findOne
      summary: Get single category by id
      description: Get single category by id
      parameters:
        - name: id
          required: true
          in: path
          description: Category id, format UUID v4
          schema: {}
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              example:
                id: f3d2f4c6-5376-48df-b4a7-9fe825559db9
                name: string
                description: string
        '400':
          description: Bad request. Category id is invalid (not uuid)
        '403':
          description: Insufficient permissions
        '404':
          description: Category not found
      tags:
        - Categories
    put:
      operationId: CategoryController_update
      summary: Update category information
      description: Updates category information by ID
      parameters:
        - name: id
          required: true
          in: path
          description: Category id, format UUID v4
          schema: {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCategoryDto'
      responses:
        '200':
          description: The category has been updated
          content:
            application/json:
              example:
                id: f3d2f4c6-5376-48df-b4a7-9fe825559db9
                name: string
                description: string
        '400':
          description: Bad request. Category id is invalid (not uuid)
        '403':
          description: Insufficient permissions
        '404':
          description: Category not found
      tags:
        - Categories
    delete:
      operationId: CategoryController_remove
      summary: Delete category
      description: Deletes category. Sets categoryId to null on associated articles.
      parameters:
        - name: id
          required: true
          in: path
          description: Category id, format UUID v4
          schema: {}
      responses:
        '204':
          description: The category has been deleted
          example: ''
        '400':
          description: Bad request. Category id is invalid (not uuid)
        '403':
          description: Insufficient permissions
        '404':
          description: Category not found
      tags:
        - Categories
  /comment:
    post:
      operationId: CommentController_create
      summary: Add new comment
      description: Add a comment to an article
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentDto'
      responses:
        '201':
          description: Comment is created
          content:
            application/json:
              example:
                id: 63d364f4-168b-416f-b9d2-865d5a579e87
                content: Great article!
                articleId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                authorId: b1b73593-2445-421a-af42-359114d6c536
                createdAt: 1775287579998
        '400':
          description: Bad request. Body does not contain required fields
        '403':
          description: Insufficient permissions
        '422':
          description: Unprocessable entity. Referenced articleId doesn't exist
      tags:
        - Comments
    get:
      operationId: CommentController_findAll
      summary: Get comments for an article
      description: >-
        Gets all comments for a specific article. Requires articleId query
        parameter.
      parameters:
        - name: page
          required: false
          in: query
          description: Page number (starting from 1)
          schema:
            example: 1
            type: number
        - name: limit
          required: false
          in: query
          description: Number of items per page
          schema:
            example: 10
            type: number
        - name: articleId
          required: true
          in: query
          description: Article ID to get comments for
          schema:
            example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
            type: string
        - name: sortBy
          required: false
          in: query
          description: The field to sort by
          schema:
            example: content
            enum:
              - id
              - content
              - authorId
              - articleId
              - createdAt
            type: string
        - name: order
          required: false
          in: query
          description: Sorting direction
          schema:
            example: DESC
            enum:
              - ASC
              - DESC
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              examples:
                Without pagination:
                  value:
                    - id: 63d364f4-168b-416f-b9d2-865d5a579e87
                      content: Great article!
                      articleId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                      authorId: b1b73593-2445-421a-af42-359114d6c536
                      createdAt: 1775287579998
                  description: Without page and limit query parameters
                With pagination:
                  value:
                    total: 1
                    page: 1
                    limit: 10
                    data:
                      - id: 63d364f4-168b-416f-b9d2-865d5a579e87
                        content: Great article!
                        articleId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                        authorId: b1b73593-2445-421a-af42-359114d6c536
                        createdAt: 1775287579998
                  description: With page and limit query parameters
        '400':
          description: Bad request. Wrong query parameters (hacker scope)
        '403':
          description: Insufficient permissions
      tags:
        - Comments
  /comment/{id}:
    get:
      operationId: CommentController_findById
      summary: Get single comment by id
      description: Gets single comment by id
      parameters:
        - name: id
          required: true
          in: path
          description: Comment id, format UUID v4
          schema: {}
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              example:
                id: 63d364f4-168b-416f-b9d2-865d5a579e87
                content: Great article!
                articleId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                authorId: b1b73593-2445-421a-af42-359114d6c536
                createdAt: 1775287579998
        '400':
          description: Bad request. Comment id is invalid (not uuid)
        '403':
          description: Insufficient permissions
        '404':
          description: Comment not found
      tags:
        - Comments
    delete:
      operationId: CommentController_remove
      summary: Delete comment
      description: Deletes comment
      parameters:
        - name: id
          required: true
          in: path
          description: Comment id, format UUID v4
          schema: {}
      responses:
        '204':
          description: The comment has been deleted
          example: ''
        '400':
          description: Bad request. Comment id is invalid (not uuid)
        '403':
          description: Insufficient permissions
        '404':
          description: Comment not found
      tags:
        - Comments
...
info:
  title: Knowledge Hub
  description: Knowledge hub service for managing articles, categories, and comments
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    CreateCategoryDto:
      type: object
      properties:
        name:
          type: string
          example: string
        description:
          type: string
          example: string
      required:
        - name
        - description
    UpdateCategoryDto:
      type: object
      properties:
        name:
          type: string
          example: string
        description:
          type: string
          example: string
    CreateCommentDto:
      type: object
      properties:
        content:
          type: string
          example: string
        articleId:
          type: string
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        authorId:
          type: string
          example: b1b73593-2445-421a-af42-359114d6c536
      required:
        - content
        - articleId
        - authorId
...